dom/tests/mochitest/mouselock/test_constantxy.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=92264
-->
<head>
  <title>Test for Bug 633602</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="runTests();">
   <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
   <p id="display"></p>
   <div id="content" style="display: none">
   </div>
   <canvas style="background-color: #CCCCCC;" id="canvas" width="150" height="150"></canvas>
   <pre id="test">
      <script>
         SimpleTest.waitForExplicitFinish();
         var canvas = document.getElementById("canvas");
         var pointer = navigator.pointer;
         var move = 0;
         var clientX1, clientY1;
         var screenX1, screenY1;
         SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);

         // When fullscreen is enabled on canvas - lock the mouse 
         document.addEventListener('mozfullscreenchange', function() {
            if(document.mozFullScreen) {
               if(document.mozFullScreenElement == canvas) {
                 pointer.lock(canvas, function() {
                   canvas.addEventListener('mousemove', function(e) {
                     if(move === 0) {
                       clientX1 = e.clientX;
                       clientY1 = e.clientY;
                       screenX1 = e.screenX;
                       screenY1 = e.screenY;
                       move++;
                       synthesizeMouse(canvas, 15, 15, {type: "mousemove"}, window);
                     } else if (move === 1) {
                       is(clientX1, e.clientX, "clientX must remain constant");
                       is(clientY1, e.clientY, "clientY must remain constant");
                       is(screenX1, e.screenX, "screenX must remain constant");
                       is(screenY1, e.screenY, "screenY must remain constant");
                       move++;
                       synthesizeMouse(canvas, 20, 20, {type: "mousemove"}, window);
                     } else if(move === 2) {   
                       SimpleTest.finish();
                       document.mozCancelFullScreen();
                     }
                   }, false);
                   synthesizeMouse(canvas, 10, 10, {type: "mousemove"}, window);
                 });
               }
             }
           }, false);
           function runTests() {
             canvas.mozRequestFullScreen();
           }
     </script>
   </pre>
</body>
</html>
